feat: allow env0_environment creation to be synchronous
This commit adds a new "wait" parameter to the env0_environment
resource. If this parameter is set to true, the resource waits for the
environment's latest deployment to hit a terminal state. This allows
multiple environments to be chained together with terraform
dependencies to deploy environments in a specific order. The parameter
defaults to false to preserve existing behaviour